Mejorar manejador de errores (#21)
* se simplifica logica de mostrar errores

* fix al import de lodash
